refactor: relocate NetworkToken card component
